# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/common/report": {
            "post": {
                "description": "使用但不校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "埋点上报",
                "parameters": [
                    {
                        "description": "token和埋点数据",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.ReportPayload"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "ok"
                    }
                }
            }
        },
        "/common/statics": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取统计数据",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "end_timestamp",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "start_timestamp",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"统计数据",
                        "schema": {
                            "$ref": "#/definitions/entity.GetStaticsResp"
                        }
                    }
                }
            }
        },
        "/doc/create": {
            "post": {
                "description": "校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "新建文档",
                "parameters": [
                    {
                        "description": "空间ID、标题、封面、内容",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.CreateDocPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"文档详情",
                        "schema": {
                            "$ref": "#/definitions/model.Doc"
                        }
                    }
                }
            }
        },
        "/doc/delete": {
            "post": {
                "description": "校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "删除文档",
                "parameters": [
                    {
                        "description": "文档ID列表",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.DeleteDocPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            }
        },
        "/doc/docs": {
            "get": {
                "description": "使用但不校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "搜索文档",
                "parameters": [
                    {
                        "type": "string",
                        "name": "author_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "keywords",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sort_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "space_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"文档列表",
                        "schema": {
                            "$ref": "#/definitions/entity.GetDocsResp"
                        }
                    }
                }
            }
        },
        "/doc/drafts": {
            "get": {
                "description": "校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "获取文档草稿",
                "parameters": [
                    {
                        "type": "string",
                        "name": "doc_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"草稿列表",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Draft"
                            }
                        }
                    }
                }
            }
        },
        "/doc/info": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取文档详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文档ID",
                        "name": "doc_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"文档详情",
                        "schema": {
                            "$ref": "#/definitions/entity.GetDocInfoResp"
                        }
                    }
                }
            }
        },
        "/doc/update": {
            "post": {
                "description": "校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "更新文档",
                "parameters": [
                    {
                        "description": "文档ID、空间ID、标题、封面、内容、摘要",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.UpdateDocPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"更新后的文档详情",
                        "schema": {
                            "$ref": "#/definitions/model.Doc"
                        }
                    }
                }
            }
        },
        "/doc/update_drafts": {
            "post": {
                "description": "校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "更新文档草稿",
                "parameters": [
                    {
                        "description": "文档ID和草稿内容",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.UpdateDraftPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"更新后的草稿详情",
                        "schema": {
                            "$ref": "#/definitions/model.Draft"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.Author": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                }
            }
        },
...
        "entity.CreateDocPayload": {
            "type": "object",
            "required": [
                "space_id",
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "cover": {
                    "type": "string"
                },
                "space_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "entity.DeleteDocPayload": {
            "type": "object",
            "required": [
                "doc_ids"
            ],
            "properties": {
                "doc_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "entity.DocInfo": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/entity.Author"
                },
                "author_id": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "cover": {
                    "type": "string"
                },
                "creation_time": {
                    "type": "string"
                },
                "doc_id": {
                    "type": "string"
                },
                "public": {
                    "type": "boolean"
                },
                "space_id": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "update_time": {
                    "type": "string"
                }
            }
        },
...
        "entity.GetDocInfoResp": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/entity.Author"
                },
                "author_id": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "cover": {
                    "type": "string"
                },
                "creation_time": {
                    "type": "string"
                },
                "doc_id": {
                    "type": "string"
                },
                "public": {
                    "type": "boolean"
                },
                "space_id": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "update_time": {
                    "type": "string"
                }
            }
        },
        "entity.GetDocsResp": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.DocInfo"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "entity.GetStaticsResp": {
            "type": "object",
            "properties": {
                "pv": {
                    "type": "integer"
                },
                "uv": {
                    "type": "integer"
                }
            }
        },
...
        "entity.ReportPayload": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {}
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "entity.UpdateDocPayload": {
            "type": "object",
            "required": [
                "doc_id"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "cover": {
                    "type": "string"
                },
                "doc_id": {
                    "type": "string"
                },
                "public": {
                    "type": "boolean"
                },
                "summary": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "entity.UpdateDraftPayload": {
            "type": "object",
            "required": [
                "content",
                "doc_id"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "doc_id": {
                    "type": "string"
                }
            }
        },
...
        "model.Doc": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "cover": {
                    "type": "string"
                },
                "creation_time": {
                    "type": "string"
                },
                "doc_id": {
                    "type": "string"
                },
                "public": {
                    "type": "boolean"
                },
                "space_id": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "update_time": {
                    "type": "string"
                }
            }
        },
        "model.Draft": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "creation_time": {
                    "type": "string"
                }
            }
        },
...
    }
}